feat: Added a loading bar when you post a new question + added component shell for InteractivePollItem.js
Used react-loading-bar package and had to figure out how to use it. You had to mount it for it to show on your UI. Then you have to import and use the prewritten reducer as well as action creator. Regarding InteractivePollItem.js, had trouble accessing the question id cos there was no props to grab it from. But learnt that you can actually grab data off of the URL itself via the React Router match and location props which you get for free by just using React Router. So I just had to go this.props.match.params.id and voila that solved my problem