feat: Implemented InteractivePollItem.js Poll Stats functionality
Whereas in the previous commit I implemented the Vote functionality of InteractivePollItem.js now I added the Vote functionality. Basically this is the same component but with 2 UI views: if the question hasn't been answered by the user yet it will show as a form where the user can vote and submit their vote. But if the user has already answered this question before it will render a different UI view: that shows the stats of how many users voted and what they selected and the percentage of votes each way. Along the way also added bare bones UI shell for Notfound.js component. Also did a couple of fixes: one was making some errors in the root reducer that wasn't updating the votes correctly. Instead of appending to the array it was replacing it completely. Finally fixed an issue whereby as soon as a user submits a vote, for a split second the view would change to the Stats view before rerouting to the Dashboard. So it looks really weird. To solve this was as simple as moving the dispatch call to happen after the DB call that way the Redux store never has a chance to be ahead of the DB and cause weird rendering issues