feat: Finishing the last of my styling and quite happy with it
Added a progress bar which was interesting. Was able to add in a function as the value for the style property and made that function return a JS object so that was a cool feature which let my bar render and update dynamically. Made Dashboard.js render the cards list horizontally and flexWrapped so that it overflows to the next line. Set the minHeight so that I give the cards uniformity and a better overall look. More experience with Flexbox and understanding justifyContent + alignItems