feat: Made styling responsive and mobile-friendly
This proved to be a huge pain and it was very difficult but using Flexbox it kind of gives you responsiveness out of the box but takes heaps of trial and error nevertheless. Also tried playing around with firebase to try deploying it there since thet last time I used Heroku. Will try to get some experience with Firebase